Move AdminRpcRequestMetadataPostInit to solana-core
