+++ core/src/admin_rpc_post_init.rs
@@ -1,0 +1,17 @@
+use {
+    solana_gossip::cluster_info::ClusterInfo,
+    solana_runtime::bank_forks::BankForks,
+    solana_sdk::pubkey::Pubkey,
+    std::{
+        collections::HashSet,
+        sync::{Arc, RwLock},
+    },
+};
+
+#[derive(Clone)]
+pub struct AdminRpcRequestMetadataPostInit {
+    pub cluster_info: Arc<ClusterInfo>,
+    pub bank_forks: Arc<RwLock<BankForks>>,
+    pub vote_account: Pubkey,
+    pub repair_whitelist: Arc<RwLock<HashSet<Pubkey>>>,
+}
