+++ validator/src/bin/solana-test-validator.rs
@@ -6,7 +6,9 @@ use {
         input_parsers::{pubkey_of, pubkeys_of, value_of},
         input_validators::normalize_to_url_if_moniker,
     },
-    solana_core::tower_storage::FileTowerStorage,
+    solana_core::{
+        admin_rpc_post_init::AdminRpcRequestMetadataPostInit, tower_storage::FileTowerStorage,
+    },
     solana_faucet::faucet::run_local_faucet_with_port,
     solana_rpc::{
         rpc::{JsonRpcConfig, RpcBigtableConfig},
@@ -561,13 +563,12 @@ fn main() {
 
     match genesis.start_with_mint_address(mint_address, socket_addr_space) {
         Ok(test_validator) => {
-            *admin_service_post_init.write().unwrap() =
-                Some(admin_rpc_service::AdminRpcRequestMetadataPostInit {
-                    bank_forks: test_validator.bank_forks(),
-                    cluster_info: test_validator.cluster_info(),
-                    vote_account: test_validator.vote_account_address(),
-                    repair_whitelist: test_validator.repair_whitelist(),
-                });
+            *admin_service_post_init.write().unwrap() = Some(AdminRpcRequestMetadataPostInit {
+                bank_forks: test_validator.bank_forks(),
+                cluster_info: test_validator.cluster_info(),
+                vote_account: test_validator.vote_account_address(),
+                repair_whitelist: test_validator.repair_whitelist(),
+            });
             if let Some(dashboard) = dashboard {
                 dashboard.run(Duration::from_millis(250));
             }
